+++ sdk/capacitor/src/index.ts
@@ -50,7 +50,7 @@ export class MobileLabHttpTransport implements MobileLabTransport {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
-          'X-MobileLab-SDK': '@mobilelab/capacitor/0.5.0',
+          'X-MobileLab-SDK': '@mobilelab/capacitor/0.6.0',
         },
         body: JSON.stringify(event),
         signal: controller.signal,